update feedback v2

+++ app/Config/Routes.php
@@ -14,56 +14,59 @@ $routes->get('/', 'Home::index');
 
 
 // router admin 
-$routes->get('/admin', 'AdminController::index');
-$routes->get('akun', 'AdminController::akun');
-$routes->get('rps', 'AdminController::rps');
+$routes->get('/admin', 'AdminController::index', ['filter' => 'role:admin']);
+$routes->get('akun', 'AdminController::akun', ['filter' => 'role:admin']);
+$routes->get('rps', 'AdminController::rps', ['filter' => 'role:admin']);
 
-$routes->get('tambah', 'AdminController::tambah');
-$routes->post('tambah/adduser', 'AdminController::adduser');
+$routes->get('tambah', 'AdminController::tambah', ['filter' => 'role:admin']);
+$routes->post('tambah/adduser', 'AdminController::adduser', ['filter' => 'role:admin']);
 
 // $routes->get('edit', 'Admin::edit');
-$routes->get('edit/(:num)', 'AdminController::edit/$1');
-$routes->post('edit/edituser/(:num)', 'AdminController::updateuser/$1');
-$routes->post('deleteuser/(:num)', 'AdminController::deleteuser/$1');
+$routes->get('edit/(:num)', 'AdminController::edit/$1', ['filter' => 'role:admin']);
+$routes->post('edit/edituser/(:num)', 'AdminController::updateuser/$1', ['filter' => 'role:admin']);
+$routes->post('deleteuser/(:num)', 'AdminController::deleteuser/$1', ['filter' => 'role:admin']);
 
-$routes->get('tambahrp', 'AdminController::tambahrp');
-$routes->post('tambahrp/addrp', 'AdminController::addrp');
+$routes->get('tambahrp', 'AdminController::tambahrp', ['filter' => 'role:admin']);
+$routes->post('tambahrp/addrp', 'AdminController::addrp', ['filter' => 'role:admin']);
 
 // $routes->get('editrp', 'Admin::editrp');
-$routes->get('editrp/(:num)', 'AdminController::editrp/$1');
-$routes->post('editrp/updaterp/(:num)', 'AdminController::updaterp/$1');
-$routes->post('editrp/deleterp/(:num)', 'AdminController::deleterp/$1');
+$routes->get('editrp/(:num)', 'AdminController::editrp/$1', ['filter' => 'role:admin']);
+$routes->post('editrp/updaterp/(:num)', 'AdminController::updaterp/$1', ['filter' => 'role:admin']);
+$routes->post('editrp/deleterp/(:num)', 'AdminController::deleterp/$1', ['filter' => 'role:admin']);
 
-$routes->get('profil', 'AdminController::profil');
-$routes->get('notif', 'AdminController::notif');
+$routes->get('profil', 'AdminController::profil', ['filter' => 'role:admin']);
+$routes->get('notif', 'AdminController::notif', ['filter' => 'role:admin']);
 // $routes->get('/dosen', 'dosen::index');
 // $routes->get('/gpm', 'gpm::index');
 // route admin selesai
 
 // route dosen
 $routes->get('/dosen', 'dosenController::dosen', ['filter' => 'role:dosen']);
-$routes->get('dosen/unggah-rps', 'dosenController::unggah_rps');
-$routes->get('dosen/linkRPS', 'dosenController::link_rps');
-$routes->post('dosen/simpan_rps', 'UnggahRpsController::simpan_rps');
-$routes->get('dosen/daftar_upload', 'dosenController::daftar_upload');
-$routes->get('dosen/isi_bap', 'dosenController::bap');
-$routes->get('dosen/daftar_bap', 'dosenController::daftar_bap');
-$routes->get('dosen/feedback', 'dosenController::feedback');
-$routes->get('dosen/getMataKuliahByJurusan/(:num)', 'dosenController::getMataKuliahByJurusan/$1');
-$routes->delete('dosen/hapus_rps/(:num)', 'dosenController::hapus_rps/$1');
-$routes->get('dosen/get_rps/(:num)', 'dosenController::get_rps/$1');
-$routes->put('dosen/update_rps/(:num)', 'dosenController::update_rps/$1');
-$routes->post('dosen/simpan_bap', 'dosenController::simpan_bap');
-$routes->get('dosen/get-bap-details/(:num)', 'Dosen::getBapDetails/$1');
-$routes->post('/dosen/update-bap', 'Dosen::updateBap');
-$routes->get('/dosen/download-bap/(:num)', 'Dosen::downloadBap/$1');
-$routes->delete('/dosen/delete-bap/(:num)', 'Dosen::deleteBap/$1');
-$routes->get('dosen/profile', 'dosenController::profile_dosen');
-$routes->get('dosen/notifikasi_rps', 'dosenController::notifikasi_rps');
+$routes->get('dosen/unggah-rps', 'dosenController::unggah_rps', ['filter' => 'role:dosen']);
+$routes->get('dosen/linkRPS', 'dosenController::link_rps', ['filter' => 'role:dosen']);
+$routes->post('dosen/simpan_rps', 'UnggahRpsController::simpan_rps', ['filter' => 'role:dosen']);
+$routes->get('dosen/daftar_upload', 'dosenController::daftar_upload', ['filter' => 'role:dosen']);
+$routes->get('dosen/isi_bap', 'dosenController::bap', ['filter' => 'role:dosen']);
+$routes->get('dosen/daftar_bap', 'dosenController::daftar_bap', ['filter' => 'role:dosen']);
+$routes->get('dosen/feedback', 'dosenController::feedback', ['filter' => 'role:dosen']);
+$routes->get('dosen/getMataKuliahByJurusan/(:num)', 'dosenController::getMataKuliahByJurusan/$1', ['filter' => 'role:dosen']);
+$routes->delete('dosen/hapus_rps/(:num)', 'dosenController::hapus_rps/$1', ['filter' => 'role:dosen']);
+$routes->get('dosen/get_rps/(:num)', 'dosenController::get_rps/$1', ['filter' => 'role:dosen']);
+$routes->put('dosen/update_rps/(:num)', 'dosenController::update_rps/$1', ['filter' => 'role:dosen']);
+$routes->post('dosen/simpan_bap', 'dosenController::simpan_bap', ['filter' => 'role:dosen']);
+$routes->get('dosen/get-bap-details/(:num)', 'dosenController::getBapDetails/$1', ['filter' => 'role:dosen']);
+$routes->post('/dosen/update-bap', 'dosenController::updateBap', ['filter' => 'role:dosen']);
+$routes->get('/dosen/download-bap/(:num)', 'dosenController::downloadBap/$1', ['filter' => 'role:dosen']);
+$routes->delete('/dosen/delete-bap/(:num)', 'dosenController::deleteBap/$1', ['filter' => 'role:dosen']);
+$routes->get('dosen/profile', 'dosenController::profile_dosen', ['filter' => 'role:dosen']);
+$routes->get('dosen/notifikasi_rps', 'dosenController::notifikasi_rps', ['filter' => 'role:dosen']);
 
 
-// route gpm
-$routes->get('/dashboard/dashboard_gpm', 'GpmController::dashboard_gpm', ['filter' => 'role:gpm']);
-$routes->get('/dashboard/gpm_rps', 'GpmController::gpm_rps', ['filter' => 'role:gpm']);
-$routes->get('gpm/download/(:num)', 'GpmController::download/$1', ['filter' => 'role:gpm']);
-$routes->post('gpm/save-review', 'GpmController::saveReview', ['filter' => 'role:gpm']);
+
+// route gpm dan kajur
+
+$routes->get('/kajur', 'GpmController::dashboard_kajur', ['filter' => 'role:gpm,kajur']);
+$routes->get('/gpm', 'GpmController::dashboard_gpm', ['filter' => 'role:gpm,kajur']);
+$routes->get('/dashboard/gpm_rps', 'GpmController::gpm_rps', ['filter' => 'role:gpm,kajur']);
+$routes->get('gpm/download/(:num)', 'GpmController::download/$1', ['filter' => 'role:gpm,kajur']);
+$routes->post('gpm/save-review', 'GpmController::saveReview', ['filter' => 'role:gpm,kajur']);
